Move buildins to package plugin-javascript
