+++ packages/core/package.json
@@ -8,7 +8,6 @@
   "dependencies": {
     "@octolinker/blob-reader": "1.0.0",
     "@octolinker/helper-settings": "1.0.0",
-    "builtins": "^1.0.3",
     "github-injection": "^1.0.1",
     "jquery": "^3.2.1",
     "querystring": "^0.2.0",
