+++ packages/plugin-javascript/builtins-docs.js
